backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
scikit-learn>=1.5.0
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.4.0
tldextract>=5.1.0
python-multipart>=0.0.9
pydantic>=2.7.0
aiofiles>=23.2.0
python-dotenv>=1.0.0
google-genai>=0.2.2
python-whois==0.8.0
requests==2.31.0
PyMuPDF==1.23.8
Pillow==10.2.0
